requirements.txt
PyQt5>=5.6
pycryptodomex
    